+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes of SVRUnit releases are documented in this file
 using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
-## [unreleased]
+## [1.4.0]
 
 ### Added